WIP: Compile all JS assets or use NPM modules
