+++ package.json
@@ -15,12 +15,14 @@
     "browse": "live-server dist/",
     "build": "npm-run-all --parallel copy-public webpack",
     "dev-build": "npm-run-all --parallel copy-public dev-webpack",
-    "start": "cross-env NODE_ENV=development webpack serve --mode development --open",
+    "server": "cross-env NODE_ENV=development webpack serve --mode development --open",
+    "start": "npm-run-all copy-public server",
     "dev-start": "npm-run-all --parallel copy-public copy-public:watch dev-webpack:watch browse",
     "deploy": "./build_and_deploy.sh",
     "cloudfront:invalidations": "s3cmd cfinvalinfo cf://E3V22UD5MXF4CR",
     "cloudfront:inval-root": "aws cloudfront create-invalidation --distribution-id E3V22UD5MXF4CR --paths /index.html",
     "test:full": "npm-run-all test test:cypress",
+    "test:cypress:start": "npm-run-all copy-public server",
     "test:cypress": "cypress run --config video=false,defaultCommandTimeout=10000 --env testEnv=local",
     "test:cypress:ci": "cypress run --config defaultCommandTimeout=10000 --env testEnv=local --record",
     "test:cypress:smoke": "cypress run --spec 'cypress/integration/smoke/*' --config video=false,defaultCommandTimeout=10000 --env testEnv=local",
