+++ src/public/index.html
@@ -14,10 +14,6 @@
   </head>
   <body>
     <div id="app"></div>
-
-    <script src="./assets/js/globals.js"></script>
-    <script src="./assets/js/iframe-phone.js"></script>
-    <script src="./assets/js/CodapInterface.js"></script>
     <script src="./assets/js/app.js"></script>
   </body>
 </html>
